refactor processing layer, create task context
