+++ backend/apps/ifc_validation/tasks/logger.py
@@ -1,0 +1,2 @@
+from celery.utils.log import get_task_logger
+logger = get_task_logger("ifc_validation")
